+++ src/config/http.rs
@@ -1,22 +1,20 @@
-/*
- * This file is part of Astarte.
- *
- * Copyright 2022 SECO Mind Srl
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
+// This file is part of Astarte.
+//
+// Copyright 2022, 2026 SECO Mind Srl
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
 
 //! Provides an HTTP API to set The Message Hub configurations
 
@@ -33,8 +31,8 @@ use axum::{Json, Router};
 use log::debug;
 use serde::{Deserialize, Serialize};
 use tokio::net::TcpListener;
-use tokio::sync::mpsc::error::SendError;
 use tokio::sync::Notify;
+use tokio::sync::mpsc::error::SendError;
 use tokio::task::{JoinError, JoinHandle};
 use tokio_util::sync::CancellationToken;
 
@@ -258,6 +256,8 @@ mod test {
     #[tokio::test]
     #[serial]
     async fn server_test() {
+        let _ = rustls::crypto::aws_lc_rs::default_provider().install_default();
+
         let notify = Arc::new(Notify::new());
 
         let dir = TempDir::new().unwrap();
@@ -312,6 +312,8 @@ mod test {
     #[tokio::test]
     #[serial]
     async fn bad_request_test() {
+        let _ = rustls::crypto::aws_lc_rs::default_provider().install_default();
+
         let notify = Arc::new(Notify::new());
 
         let dir = TempDir::new().unwrap();
@@ -348,6 +350,8 @@ mod test {
     #[tokio::test]
     #[serial]
     async fn test_set_config_invalid_cfg() {
+        let _ = rustls::crypto::aws_lc_rs::default_provider().install_default();
+
         let notify = Arc::new(Notify::new());
 
         let dir = TempDir::new().unwrap();