+++ CHANGELOG.md
@@ -1,9 +1,36 @@
+<!--
+This file is part of Astarte.
+
+Copyright 2023-2026 SECO Mind Srl
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+   http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+SPDX-License-Identifier: Apache-2.0
+-->
+
 # Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Changed
+
+- Update the MSRV to 1.86 and edition 2024
+  [#430](https://github.com/astarte-platform/astarte-message-hub/pull/430)
 
 ## [0.9.2] - 2026-01-29
 
@@ -73,13 +100,13 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Let a node receive only the properties in its introspection
   [#335](https://github.com/astarte-platform/astarte-message-hub/pull/335).
 
-
 ## [0.7.4] - 2025-07-02
 
 ### Added
 
-- Introduce feature `vendored`, `bindgen` and `cross` to vendor, use `aws_lc` as default crypto provider
-  and cross compile [#363](https://github.com/astarte-platform/astarte-message-hub/pull/363).
+- Introduce feature `vendored`, `bindgen` and `cross` to vendor, use `aws_lc` as default crypto
+  provider and cross compile
+  [#363](https://github.com/astarte-platform/astarte-message-hub/pull/363).
 - Implement methods to retrieve Astarte properties
   [#329](https://github.com/astarte-platform/astarte-message-hub/pull/329).